\name{optionalMatrix-class}
\Rdversion{1.1}
\docType{class}
\alias{optionalMatrix-class}

\title{Class optionalMatrix}
\description{Class optionalMatrix.}
\section{Objects from the Class}{
  A virtual Class: No objects may be created from it.
}
\section{Methods}{
No methods defined with class "optionalMatrix" in the signature.
}
\author{Georgi N. Boshnakov}
\examples{
showClass("optionalMatrix")
}
\keyword{classes}
\keyword{internal}
